CUDASIMDTypes.jl

CUDASIMDTypes.bitifelse — Method

bitifelse(cond, x, y)

Bitwise version of ifelse.

For each bit of the output, the respective bit in cond determines whether the respective bit of x or of y is selected.

CUDASIMDTypes.lop3 — Method

lop3(a, b, c, lut)

Arbitrary logical operation on 3 inputs.

Call the PTX prmt instruction. This computes a bitwise logical operation on the inputs a, b, and c.

See make_lop3_lut for creating the look-up table lut.

CUDASIMDTypes.make_lop3_lut — Method

make_lop3_lut(f)

Create a look-up table for lop3.

CUDASIMDTypes.prmt — Method

prmt(a, b, op)

Permute bytes bytes from a pair of inputs.

Call the PTX prmt instruction. This picks four arbitrary bytes from the input values a and b.